code/client/requirements.txt
django-environ==0.4.5
djangorestframework==3.11.0
openpyxl==3.0.3
pexpect==4.8
scp
paramiko
django-s3-storage
troposphere==2.7.0